langchain-mcp-test/pyproject.toml
[project]
name = "langchain-mcp-test"
version = "0.1.0"
description = "Langchain MCP Test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "isort>=6.0.1",
    "google-generativeai>=0.4.1",
    "langchain-mcp-adapters>=0.1.9",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
    "langchain-google-genai>=2.0.10",
]
